Make plugin hot-reloadable
This allows the plugin to be (re)loaded without restarting the IDE,
which is useful when developing the plugin and more convenient when
upgrading.

There appear to be some issues in IntelliJ IDEA with dynamic unloading
(IDEA-289241, IDEA-289243), but in the worst case, the user receives a
message that they have to restart IDE to reload. Which was the norm
previously.

Docs: https://plugins.jetbrains.com/docs/intellij/dynamic-plugins.html